refactor: extract infra to a specific folder
