+++ app/(docs)/docs/layout.tsx
@@ -1,8 +1,7 @@
 import { GithubInfo } from "fumadocs-ui/components/github-info";
 import { DocsLayout, type DocsLayoutProps } from "fumadocs-ui/layouts/docs";
 
-import { baseOptions } from "@/integrations/fumadocs/layout.shared";
-import { source } from "@/integrations/fumadocs/source";
+import { baseOptions, source } from "@/config/docs";
 
 const docsOptions: DocsLayoutProps = {
   ...baseOptions(),
