+++ content/docs/appendix/design-decision/001-flights-booking.mdx
@@ -322,4 +322,3 @@ await db.update(flightSeatClasses).set({
 - 第 1 步结束：创建订单 + 锁定座位 + 开始倒计时
 - 第 3 步结束：支付成功 + 订单确认
 - 15 分钟后：自动取消 + 释放座位（如果未支付）
-
