+++ content/docs/contributing/coding-standards.mdx
@@ -21,7 +21,7 @@ src/
 ├── config/                 # 环境变量、UI/错误/系统常量配置
 ├── db/                     # Drizzle client/schema
 ├── services/               # 跨领域服务门面
-├── integrations/           # 外部服务适配层
+├── infra/                  # 外部服务适配层（认证/通讯/日志/HTTP/后台/文档）
 ├── lib/                    # 通用工具/封装（format/validation/server/security 等分组）
 ├── hooks/                  # React Hooks
 ├── types/                  # TypeScript 类型定义
