+++ src/domains/auth/index.ts
@@ -8,4 +8,4 @@ const authClient = createAuthClient({
 
 export const { signIn, signUp, useSession } = authClient;
 
-export { auth } from "./infra/auth.plugin";
+export { auth } from "@/infra/auth";
