+++ src/infra/communications/index.ts
@@ -1,0 +1,8 @@
+import "server-only";
+
+export { sendSmsOtp } from "./aliyun-sms.client";
+export {
+  ResendEmailClient,
+  sendEmailOtp,
+  sendOrderConfirmationEmail,
+} from "./resend.client";
