+++ src/infra/communications/resend.client.tsx
@@ -1,3 +1,5 @@
+import "server-only";
+
 import { Resend } from "resend";
 
 import { OrderConfirmationEmail, OtpEmailTemplate } from "@/components/emails";
