+++ src/infra/logging/logger.ts
@@ -1,3 +1,5 @@
+import "server-only";
+
 import pino from "pino";
 import pretty from "pino-pretty";
