+++ src/infra/notifications/index.ts
@@ -1,0 +1,4 @@
+import "server-only";
+
+export { registerAuthNotificationHandlers } from "./auth-notification.handler";
+export { sendOrderConfirmationEmail } from "./order-email-sender";
