+++ tests/setup/global.ts
@@ -1,5 +1,10 @@
 import "@testing-library/jest-dom/vitest";
 
+import { vi } from "vitest";
+
+// Mock server-only to allow testing server modules in Vitest without Next.js bundler
+vi.mock("server-only", () => ({}));
+
 // Polyfill for ResizeObserver (needed for Radix UI components)
 if (typeof ResizeObserver === "undefined") {
   globalThis.ResizeObserver = class ResizeObserver {
